SpreadsheetLocaleComponent dependency: SpreadsheetSuggestBoxComponentSuggestionsProvider
